+++ src/Foundry/Factory/AliasedFactory.php
@@ -8,14 +8,15 @@ use Zenstruck\Foundry\Persistence\Proxy;
 
 /**
  * @template T of object
+ *
  * @extends Factory<T&Proxy<T>>
  */
 abstract class AliasedFactory extends Factory
 {
+    abstract protected static function getEntityIdentifier(): string;
+
     public static function class(): string
     {
         return AliasedFactoryHelper::getClass(static::getEntityIdentifier());
     }
-
-    abstract protected static function getEntityIdentifier(): string;
 }