+++ src/Foundry/Randomizer/Randomizer.php
@@ -14,27 +14,29 @@ final class Randomizer
         $this->defaultNullProbability = $defaultNullProbability;
     }
 
-    private function validateProbability(float $probability): void
-    {
-        if ($probability < 0 || $probability > 1) {
-            throw new \InvalidArgumentException('Probability must be between 0 and 1');
-        }
-    }
-
     /**
      * @template V
+     *
      * @param V $value
      *
      * @return V|null
      */
     public function valueOrNull(mixed $value, ?float $nullProbability = null): mixed
     {
-        if ($nullProbability !== null) {
+        if (null !== $nullProbability) {
             $this->validateProbability($nullProbability);
         }
+
         $nullProbability ??= $this->defaultNullProbability;
         $randomFloat = mt_rand() / (mt_getrandmax() + 1);
 
-        return ($randomFloat < $nullProbability) ? null : $value;
+        return $randomFloat < $nullProbability ? null : $value;
+    }
+
+    private function validateProbability(float $probability): void
+    {
+        if ($probability < 0 || $probability > 1) {
+            throw new \InvalidArgumentException('Probability must be between 0 and 1');
+        }
     }
 }